server/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
filelock==3.6.0
Flask==2.0.3
flatbuffers==2.0
gast==0.5.3
google-generativeai==0.7.1
google-pasta==0.2.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
networkx==2.6.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pymongo==4.0.2
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scipy==1.8.0
sentencepiece==0.1.96
six==1.16.0
SpeechRecognition==3.8.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
